server/requirements.txt
alembic==1.13.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
Babel==2.15.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
beniget==0.4.1
bidict==0.23.1
blinker==1.6.2
Bottleneck==1.3.2
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
ccsm==0.9.14.1
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.2.0
chrome-gnome-shell==0.0.0
click==8.1.6
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
constantly==23.10.4
contourpy==1.1.0
cryptography==43.0.0
cupshelpers==1.0
cycler==0.11.0
daphne==4.1.2
dbus-python==1.2.18
decorator==5.1.1
deep-translator==1.11.4
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-allauth==0.47.0
django-cors-headers==4.4.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-filter==24.2
django-redis==5.4.0
django-rest-framework==0.1.0
django-selenium==0.9.8
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
Flask==2.3.2
flask-babel==4.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.41.1
fs==2.4.12
future==0.18.2
gast==0.5.2
googletrans==4.0.0rc1
greenlet==3.0.3
gTTS==2.5.1
gunicorn==22.0.0
gyp==0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.0.1
incremental==24.7.2
inflect==7.2.1
iniconfig==1.1.1
ipython==8.24.0
itsdangerous==2.1.2
jdcal==1.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jsonpickle==3.0.4
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libretranslatepy==2.1.1
libvirt-python==8.0.0
llvmlite==0.38.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.3.2
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.8
netifaces==0.11.0
numba==0.55.1
numexpr==2.8.1
numpy==1.24.4
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
parso==0.8.4
patsy==0.5.3
pexpect==4.9.0
Pillow==10.0.0
playsound==1.3.0
pluggy==0.13.0
ply==3.11
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.0.9
pyRFC3339==1.1
PySocks==1.7.1
pytest==6.2.5
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
python3-openid==3.2.0
pythran==0.10.0
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.8
requests==2.31.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.20.0
service-identity==24.1.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.29
sqlparse==0.4.2
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.0
stripe==10.4.0
supervisor==4.2.5
sympy==1.9
systemd-python==234
tables==3.7.0
toml==0.10.2
tomli==2.0.1
traitlets==5.14.3
translate==3.6.1
translator==0.0.9
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typeguard==4.3.0
typing_extensions==4.10.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.0.4
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==2.3.6
word2number==1.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==3.16.2
zope.interface==7.0.1
